+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-"use client"
+// "use client"
 import { Box, createTheme, ThemeProvider } from "@mui/material";
 import Image from "next/image";
 import { FilledButton } from "@/app/components/buttons/filled-button";
@@ -11,35 +11,7 @@ import { signIn, useSession } from "next-auth/react";
 import { getServerSession } from "next-auth";
 import LogOutIcon from "./components/icons/logout-icon";
 
-const darkTheme = createTheme({
-  palette: {
-    mode: "dark",
-    primary: {
-      main: '#FFA833'
-    },
-    secondary: {
-      main: '#242424',
-      light: '#3F3F3F',
-      dark: '#121212'
-    }
-  }
-})
-
 export default async function Home() {
-  // const router = useRouter();
-  //const session = getServerSession();
-  // const { data: session } = useSession()
-
-  // useEffect(() => {
-  //   const checkSession = () => {
-  //     if (session) {
-  //       router.push('/pages/Dasboard')
-  //     } else {
-  //       router.push('/pages/Login')
-  //     }
-  //   }
-  //   checkSession();
-  // }, [router]);
   const session = await getServerSession();
   return (
     <main>
